de_mcmc

Bayesian inference for a deterministic DE model (with models solved via an DE solver) with an observation model.

de_mcmc(
  N,
  data,
  de.model,
  obs.model,
  all.params,
  ref.params = NULL,
  ref.inits = NULL,
  Tmax,
  data.times,
  cnt = 10,
  plot = TRUE,
  sizestep = 0.01,
  solver = "ode",
  verbose.mcmc = TRUE,
  verbose = FALSE,
  ...
)

Arguments

N
integer, number of MCMC iterations

data
data.frame of time course observations to fit the model to. The observations must be ordered ascending by time.

de.model
a function defining a DE model, compliant with the solvers in deSolve or PBSddesolve

obs.model
a function defining an observation model. Must be a function with arguments 'data', 'sim.data', 'samp'.

all.params
debinfer_parlist containing all model, MCMC, and observation

ref.params
an optional named vector containing a set of reference parameters, e.g. the true parameters underlying a simulated data set

ref.inits
an optional named vector containing a set of reference initial values, e.g. the true initial values underlying a simulated data set

Tmax
maximum timestep for solver

data.times
time points for which observations are available

cnt
integer interval at which to print and possibly plot information on the current state of the MCMC chain

plot
logical, plot traces for all parameters at the interval defined by cnt

sizestep
timestep for solver to return values at, only used if data.times is missing

solver
the solver to use. 1 or "ode" = deSolve::ode; 2 or "dde" = PBSddesolve::dde; 3 or "dede" = deSolve::dde

verbose.mcmc
logical display MCMC progress messages

verbose
logical display verbose solver output

...
further arguments to the solver

Value

a debinfer_result object containing input parameters, data and MCMC samples