post_sim

post_sim

post_sim(x, n = 100, times, output = "all", burnin = NULL, prob = 0.95, ...)

Arguments

x
debinfer_result object

n
number of simulations

times
numeric a vector of times at which the ODE is to be evaluated. Defaults to NULL.

output
character, "sims", "all", "HDI"

burnin
integer, number of samples to discard from the start of the mcmc chain

prob
A numeric scalar in the interval (0,1) giving the target probability content of the intervals. The nominal probability content of the intervals is the multiple of 1/nrow(obj) nearest to prob.

...
additional arguments to solver

Value

a post_sim object containing a list of de solutions or summaries thereof